abstract this component out of the ARs
