--- frontend/src/pages/ApprovedActivityReport/components/ApprovedReportSection.scss
@@ -1,41 +1,0 @@
-@use '../../../colors.scss' as *;
-
-// these next two style blocks will only apply on desktop
-.ttahub-approved-report-section--heading--section-row-data  {
-  flex-basis: max(300px, 75%)
-}
-
-.ttahub-approved-report-section--heading--section-row-title {
-  flex-basis: min(220px, 25%);
-}
-
-// a reset to squeeze those ul's in
-.ttahub-approved-report-section--heading--section-row > p > ul {
-  margin: 0;
-  list-style: none;
-  padding: 0;
-}
-
-@media screen {
-  .ttahub-approved-report-section__striped {
-    background-color: $gray-two;
-  }
-}
-
-// print styles
-
-@media print{
-  .ttahub-approved-report-section {
-    padding: 0;
-    margin-bottom: 2rem;
-    page-break-inside: avoid;
-  }
-  .ttahub-activity-report-view {
-    margin-top: 0;
-  }
-  .ttahub-activity-report-view > div.padding-x-5.padding-y-5 {
-    padding-top: 0;
-  }
-}
-
-
